refactor: 계층 분리를 컨트롤러 계층으로 통합
