+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.web.controller;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
