+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,8 +1,8 @@
-package roomescape.web.dto;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.db.domain.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationRequest(String name, LocalDate date, LocalTime time) {
